+++ samples/react-native/src/screens/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Placeholder } from '@sitecore-jss/sitecore-jss-react-native';
+import Route from './Route'
+
+const Home = ({ navigation }) => (
+  <Route 
+    path='/'
+    render={({ data }) =>
+      <Placeholder name='jss-home' rendering={data} navigation={navigation} />
+    }
+  />
+);
+
+export default Home
